pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "incident-response-env"
version = "0.1.0"
description = "Deterministic OpenEnv-style environment for SRE incident triage."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "openenv-core>=0.2.3",
  "pydantic>=2.8.0",
  "requests>=2.32.0",
  "uvicorn>=0.30.0",
]

[project.scripts]
server = "server.app:main"

[tool.setuptools.packages.find]
include = ["envs*", "incident_response_env*", "server*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
